requirements.txt
Django==5.1.2
pydantic-settings==2.4.0
django-cors-headers==4.4.0
gunicorn==22.0.0
djangorestframework==3.15.2
psycopg2-binary==2.9.9

# celery-redis
celery==5.4.0
redis!=4.5.5,<6.0.0,>=4.5.2
async-timeout>=4.0.3
